[OF-1515] build: put all WASM build artefacts in one place
- All artefacts generated by an Emscripten build are now found under:
  Tools\WrapperGenerator\Output\TypeScript\connected-spaces-platform.web
- Update documentation to reflect this.

+++ Tools/Emscripten/EmscriptenFullBuildAndConfigure.sh
@@ -27,5 +27,29 @@ fi
 
 python Tools/Emscripten/ReplaceComSpec.py
 docker run -w /src -v `pwd`:/src --rm emscripten/emsdk:$emsdk_version emmake make config="$1"_wasm clean
+if [ $? -ne 0 ]
+  then
+    cd Tools/Emscripten
+    echo ERROR: Clean failed. Have you started the docker engine?
+	exit 1
+fi
+	
 docker run -w /src -v `pwd`:/src --rm emscripten/emsdk:$emsdk_version emmake make -j 8 config="$1"_wasm
-cd Tools/Emscripten
+if [ $? -ne 0 ]
+  then
+    cd Tools/Emscripten
+    echo ERROR: Build failed. Have you started the docker engine?
+	exit 1
+fi
+
+cp -r Library/Binaries/WASM/"$1" Tools/WrapperGenerator/Output/TypeScript/connected-spaces-platform.web/
+if [ $? -ne 0 ]
+  then
+    cd Tools/Emscripten
+    echo ERROR: Copy failed.
+	exit 1
+  else
+    cd Tools/Emscripten
+    echo Success! Build located in: ..\WrapperGenerator\Output\TypeScript\connected-spaces-platform.web
+    exit 0
+fi
